Add tests to verify gui screenshots are up to date
The tests will generate gui screenshots used in the documentation.
If the screenshots no longer match the versions in the docs section
the tests will fail and update the corresponding screenshot with a new
version.

When a test fails, the new version of the screenshot must be manually
inspected to decide if it is a deliberate update of the gui, or if the
change is due to a bug in the application or test code.